requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic-settings

pandas
numpy
scipy

matplotlib
seaborn

scikit-learn

transformers
sentence-transformers
torch

gensim
sentencepiece

streamlit
